requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
mira_sdk==0.1.13
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.9.0
pynput==1.8.1
pyserial==3.5
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
